server/requirements.txt
aiohttp>=3.12.11
fastapi>=0.115.12
google-cloud-secret-manager>=2.24.0
google-genai>=1.19.0
gspread>=6.2.1
mcp>=1.9.3
numpy>=2.3.0
pydantic>=2.11.5
python-dotenv>=1.1.0
python-multipart>=0.0.20
requests==2.32.3
uvicorn>=0.34.3
google-auth==2.37.0
websockets>=15.0.1
certifi==2024.12.14
google-api-python-client==2.122.0
google-auth-oauthlib==1.2.0